requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-wtf
flask-mail
flask-jwt-extended
mysqlclient
pdfkit
python-dotenv
email_validator
